+++ lib/getCliConfig.ts
